+++ CHANGELOG.md
@@ -7,6 +7,10 @@ versioning.
 
 ## [Unreleased]
 
+### Added
+
+- Documentation translations for Telugu, Kannada, and Malayalam.
+
 ## [1.0.0] - 2026-06-07
 
 ### Added